+++ R/valueButton.R
@@ -25,15 +25,16 @@
 #'   For nested properties use `.`, e.g. `style.width` or `navigator.appName`.
 #' @param icon Icon included in button.
 #' @param width Width of the button.
+#' @param try_binding When `TRUE` and `selector` points to Shiny Binding and `attribute == "value"`
+#'   it tries to convert sourced input value using registered `inputHandler`.
 #' @param ... Extra attributes passed to `button` or `a` tag.
 #'
 #' @return A `shiny.tag` class object defining html structure of the button.
 #' @name valueButton
 #' @export
-valueButton <- function(inputId, label, selector, attribute = "value", icon = NULL, width = NULL, ...) {
+valueButton <- function(inputId, label, selector, attribute = "value", icon = NULL, width = NULL, try_binding = TRUE, ...) {
 
-  shiny::tagList(
-    value_button_dependency,
+  htmltools::attachDependencies(
     shiny::tags$button(
       id = inputId,
       style = htmltools::css(width = shiny::validateCssUnit(width)),
@@ -42,18 +43,19 @@ valueButton <- function(inputId, label, selector, attribute = "value", icon = NU
       `data-val` = NULL,
       `data-value-attribute` = attribute,
       `data-selector` = selector,
+      `data-try_binding` = try_binding,
       list(icon, label),
       ...
-    )
+    ),
+    value_button_dependency
   )
 }
 
 #' @rdname valueButton
 #' @export
-valueLink <- function(inputId, label, selector, attribute = "value", icon = NULL, ...) {
+valueLink <- function(inputId, label, selector, attribute = "value", icon = NULL, try_binding = TRUE, ...) {
 
-  shiny::tagList(
-    value_button_dependency,
+  htmltools::attachDependencies(
     shiny::tags$a(
       id = inputId,
       class = "value-button",
@@ -61,15 +63,20 @@ valueLink <- function(inputId, label, selector, attribute = "value", icon = NULL
       `data-val` = NULL,
       `data-value-attribute` = attribute,
       `data-selector` = selector,
+      `data-try_binding` = try_binding,
       list(icon, label),
       ...
-    )
+    ),
+    value_button_dependency
   )
 }
 
-value_button_dependency <- shiny::singleton(
-  shiny::tags$head(
-    shiny::tags$script(type = "text/javascript", src = "shinyGizmo/valuebutton.js")
-  )
+value_button_dependency <- htmltools::htmlDependency(
+  name = "valuebutton",
+  version = utils::packageVersion("shinyGizmo"),
+  package = "shinyGizmo",
+  src = "www",
+  script = "valuebutton.js"
 )
+
 # todo implement update method