+++ inst/www/valuebutton.js
@@ -34,8 +34,26 @@ $.extend(valueButtonBinding, {
     if (selector == "document") {
       $target = document;
     }
+    if ($(el).data("try_binding") == "TRUE") {
+      if ($(selector).hasClass('shiny-bound-input') && attribute == "value") {
+        var binding = $(selector).data('shiny-input-binding');
+        var type = binding.name;
+        if (Boolean(binding.getType)) {
+          type = binding.getType();
+        }
+        var value = binding.getValue($(selector));
+        if (type === null) {
+          return value;
+        }
+        return {"value": value, "type": type};
+      }
+    }
+
     var value = byString($target, attribute);
     return value;
+  },
+  getType: function(el) {
+    return "shinyGizmo.valuebutton";
   },
   subscribe: function(el, callback) {
     $(el).on('click.valueButtonBinding', function(event){
@@ -48,3 +66,4 @@ $.extend(valueButtonBinding, {
 });
 
 Shiny.inputBindings.register(valueButtonBinding, 'shiny.valueButtonBinding');
+Shiny.inputBindings.setPriority("shiny.valueButtonBinding", -1);